Allow galera wsrep_provider_options to be customised
This patch introduces galera_wsrep_provider_options as a list of
options instead of only catering for changing the gcache.size
option. This allows the deployer to change any number of the
wsrep_provider_options available.

Note: The variable galera_gcache_size is respected in this patch
but if galera_wsrep_provider_options is customised by a deployer
then the custom list must include reference to galera_gcache_size
for it to be respected.

DocImpact
Closes-Bug: #1464699
Co-Authored-By: Ian Cordasco <ian.cordasco@rackspace.com>
Change-Id: I78a5a1d344729fdc06636b1c6dade63a7ba59e94