Updates the container build process
The container build process needs to make sure that the service
sources are correctly setup and updated prior to running any other
playbooks.

The modification here is nessisary to break out the process for the
proxy create, apt sources deployment, the update of those sources
and keys, container upgrades and the installation of python2.7 for
use with Ansible. This also allows for better debugging of a failure
in container create as we'll now be able to tell where in the process
a failure happens and be able to use tags to resolve it.

Change-Id: I36be437303a73bbc98a1cd5297f6c65591653cd7
Closes-Bug: 1464771
(cherry picked from commit 86b6e8e6409f76c46551becdac783a76bb853980)