Generate consistent length passwords in AIO
The bootstrap-aio.sh script generates a password by reading a set number
of characters from urandom then removing those that aren't alphanumeric.
The problem with this is that its possible for 0 characters to remain.

This patch changes the order so that tr (translate) reads urandom
directly and removes non alphanumeric characters then that output is
truncated to the required length. This ensures the generated password is
always the desired length.

Change-Id: I03984ab891f88fe4b064f7f7fcb71edf48096493
Closes-Bug: #1465684