Added flag to instruct yaprt to ignore tempest
This change makes it possible for OSAD to instruct yaprt to NEVER
do anything with tempest and the requirements that are being
brought with the requirements that can cause issues with un-versioned
requirements. This flag simply instructs yaprt to index the repo
as a item in the report but in this case absolutely no requirements
will be indexed or built for tempest.

This was done because tempest does not version anything and its
forward movement makes it impossible to maintain a consistent stack
and for our purposes, we install tempest in a venv using an isolated
install so none of these tempest requirements are necessary for us
to have to deal with.

Change-Id: I619300352efe50828fc34ee1ded96e13a43522cd
Closes-Bug: #1466010
(cherry picked from commit cba5f5cf3d928da8cf25b7f2d4d2ecfc2957846f)