Move Cinder-volumes to "on metal"
Cinder-volumes is tightly coupled to the storage host and doesn't
gain from the benefits of containerization. This patch moves the default
location for cinder-scheduler into it's own container on the
"storage-infra_hosts", and the default location for cinder-volumes to be
"on metal" in the openstack_environment cinder.yml file.

Added documentation to suggest that you may want to remove "is_metal:
true" from cinder-volumes service when not using lvm.

Change-Id: I1879b1aa6dbe69a5a7e769a43b849441a38feeb8
Closes-Bug: #1433240