Added options for enabling instance_passwords
The nova.conf and tempest.conf option for `enable_instance_password` has been
added as a default. This option has a default of True in nova but false in tempest. 
This causes temptest to fail scheme validation on newer versions of temptest. 
To fix this issue the option being added with a default value of True for both
tempest and nova.

Change-Id: I19f5da9820f2367b3d8dd0a7f215aa3f3ea5f611
Partial-Bug: #1468061