Clear expired Django sessions regularly from DB
A random minute in the 21st hour UTC, daily, was chosen arbitrarily.
That it is run on a single arbitrary host from the set was done to
prevent unnecessary work being done until Bug #1424705 provides us with
a cleaner solution.

Adjusted to run once per day just to be sure that the operation doesn't
lock the table for a meaningful duration.

Making this configurable or tunable does not seem
to be necessary at this time.

Change-Id: I87164926a4d6a65d3c9f733d625a2d4af91c2597
Closes-Bug: #1466126
(cherry picked from commit 4ab949672428fda396d1d6a28149252dabc5655d)