Ensure flush-net-cache on local host
In a multiple host deployment, the deploying host (which runs the playbooks)
may not be the host for the containers being destroyed. Therefore,
flush-net-cache is not performed on the deploying host.

This would cause connectivity problem between the deploying host and the
containers before the staled net cache timed out ultimately.

Therefore, the local host shall always perform flush-net-cache when a container
is destroyed.

Change-Id: Iac0afa58084d35d285067c404de85e3466ba62e9
Closes-Bug: #1469285