Added openstack_kernel options for gc_thresh
This commit adds to the openstack_kernel_options to set the
"net.ipv4/6.neigh.default.gc_thresh*" values according to how much ram
is available on the box.

How this is being defined:
The change brings with it a filter to find the closest power of 2 from the
amount of ram discovered on the target host. If facts are disabled when the
role is called a default value of 1024 will be used. The `set_gc_val`, when
computed, has a max value of 8192. For both ipv4/6 thresh1 is half the
`set_gc_val` thresh2 is computed value, and finally both
thresh3/router.gc_thresh are double the `set_gc_val`.

The changes here should provide for a more scalable neutron networking
environment by default while also ensuring that the values computed are sane.
Additionallyi, should the user want to define their own values they can do so by
simply overriding the `set_gc_val`.

Change-Id: Ic5fd7ebdac009fa1472aeb0b0666f9b2611a31d7
Closes-Bug: #1427893
(cherry picked from commit b95827f35167ed845e66b4f7084ac95a2ff0d8aa)