Create LXC containers with a rootfs dir
Previously, an LXC container's rootfs was being placed into
/var/lib/lxc/<container>/rootfs.  Commit
https://review.openstack.org/#/c/190048/8 inadvertently changed this so
that a container's rootfs is now placed into /var/lib/lxc/<container/.
The downside here is that we now have a directory containing a
container's root fs and the LXC configuration for the container.

This change creates a new role-level variable called
lxc_container_rootfs_directory which defaults to
"{{ lxc_container_directory }}/{{ container_name }}/rootfs" and passes
this variable when creating the container.

Change-Id: Ibbaa879166f5efaf7ca96add131cc3762dd88653
Closes-Bug: #1470101