Fixes RabbitMQ guest user creation
This change removes the "guest" user. This patch fixes an issues
where the guest user was not being cared by default which could
leave a deployment vulnerable to should someone gain access to the
internal network that the messaging system runs within. To fix this
issue the guest user is simply being removed as its not needed by the
rest of the stack.

Closes-Bug: #1466982
Change-Id: I81f9295daf7923b828ad884b6707bb3f74b0684a