Fixes RabbitMQ guest user creation
This change removes the "guest" user. This patch fixes an issues
where the guest user was not being cared by default which could
leave a deployment vulnerable to should someone gain access to the
internal network that the messaging system runs within. To fix this
issue the guest user is simply being removed as its not needed by the
rest of the stack. The rpc-maas setup was updated to now support the
creation of the new monitoring user, which is being created as an
administrator

Change-Id: Ic91199abbf85a681ccb9f7eac60de1f0b582b748
Closes-Bug: #1466982