Set default inventory file for inventory-manage.py
The script inventory-manage.py takes a file argument. The default name
for this file, in the project, is openstack_inventory.json. Not only is
it the default but the rest of the project assumes that it will be that
name.

This commit sets the default argument for the file parameter to this
assumed file name - 'openstack_inventory.json'.

Change-Id: I635122d22bb0467aad21f48dccda1bc0444518c1
Closes-bug: #1454324
(cherry picked from commit 576831481262b93dfdf72f078a4b620a7ac130fb)