remove conntrackd package
As per bug discussion, conntrackd is actually not needed for netfilter
framework, and given that users do not appear to find the logs useful
(the only reason we're using conntrackd), let's just remove it.

This commit also defines a variable container_remove_packages so that
the list of packages to remove from containers can be added to in the
future.

NOTE: removing conntrackd does not unload the kernel conntrack modules
or disable the netfilter conntrack framework. It simply means we are not
gathering ostensibly useless logging. The kernel conntrack flows can
still be interrogated with the 'conntrack' userspace cli tool, for real
time debugging.

Change-Id: Ic74e65a6fe27060dc94bfc2f250cd53fb153c7c8
Closes-Bug: 1457196
(cherry picked from commit 0da48344c6bdff988473da1aa7e9e0f6ef724680)