Remove haproxy galera monitoring user
Prior to kilo, we created a static 'haproxy' MySQL user for haproxy.
In kilo and onwards, we now have a user variable called
galera_monitoring_user which defaults to monitoring.  This commit
updates scripts/run-upgrade.sh to remove the old haproxy user to
ensure we don't have a defunct MySQL user lying about.

Change-Id: I071596c4c6d881d5304fc49a7cf752d5489ee19b
Closes-Bug: #1472673