implemented RHEL support within OSAD
This is a direct port of all roles within OSAD to allow for RHEL host support
The changes bring the ability for RHEL / CentOS7 systems to be used along with
SystemD and yum package installation for base system libraries. All services
added within have been created in the exact same way as implemented when using
ubuntu which should make it possible for deployers to pick and choose a host
OS at will. I'm sure that this will need to go through lots of iteration and
before becoming mainline it will be ripped appart and reviewed in pieces.
That said this is now function and is passing tempest testing.

Change-Id: Id6e959d2d03060d3992a0effee793ec6426b1a38