Read affinity from environment
This patch allows a user to set the number of containers per host for
each type of container that has a default affinity greater than 1 for an
AIO build.

Currently this list is:
 * keystone
 * galera
 * rabbit_mq
 * horizon
 * repo

For example, one could set NUM_keystone_CONTAINERS=1 before running an
AIO to ensure that only one keystone container is built.

Change-Id: If90704a50414150ca37584e1f823ba0e85e287ac
Partially-Implements: blueprint split-aio-gates