Target AIO swift vars at specific containers
The variables defined in etc/openstack_deploy/conf.d/swift.yml are
added as hostvars by the dynamic inventory script. Currently the vars
are being added to hosts that do not require them.

There is a variable called limit_container_types that is used by the
dynamic inventory to filter the hosts to which variables are assigned.
This script makes use of that variable so that only swift hosts have
these variables set.

The dynamic inventory script checks to see if the limit_container_types
string is in the container name as the method of filtering. This means
that the AIO host, aio1, does not get swift_vars assigned to it. This
commit alters the filter so that the component is checked instead of the
container name.

Change-Id: I945594a55ce058d822b941d08eb0ad9c5614db15
Closes-bug: #1472261
(cherry picked from commit 1ef256a3082759213e41ffb136596001e88de4d2)