Add default user role for Keystone & Horizon and tasks to create it
In Keystone v3, the _member_ role is not implicitly created on first
use like it is in v2.

This patch adds variables to define the default role name for users:
 - keystone_default_role_name:
   this is the default role name from Keystone's point of view
 - horizon_default_role_name:
   this is the default user role from Horizon's point of view

Both Keystone and Horizon's tasks ensure that the role they're using
are registered in the Keystone database.

To maintain backwards compatibility the default value for both
variables is '_member_'.

DocImpact
Closes-bug: 1474916
Change-Id: Ie01e1771c0b435815dfe55fc0ba9a6d803ebe958