Remove {{ from "with_items" and "when" statements
This patch resolves a minor issue that was generating a "warning"
message when running the galera-install.yml playbook.

The "{{" and "}}" are not required for conditionals, and loop
expressions, so these can and should safely be removed from "when:"
statements and "with_items:" statements.

Functionality does not change but the warning message will go away after
this patch.

Change-Id: I3e9d3a4afb0e3d1284c78a37dba6d1534154a063
Closes-Bug: #1475328
(cherry picked from commit 57577bb369a107ed704f4791090b1c810599258c)