Set haproxy install to use latest packages
This patch changes the apt task for installing haproxy packages
from only checking for presence to always checking for the
latest package version.

This is essential to allow a deployer to switch from a
configuration that does not implement SSL to one that does.

Change-Id: Iaf6eaedba835a332920336b1cb66190924537301
Closes-Bug: #1475597
(cherry picked from commit ba7587267a24db33039948712751c3f7cd6a2eb0)