development-stack Doc Update
This patch changes how a development stack should be setup,
referencing the use of the bootstrap and run-playbook scripts
instead of the gate-check-commit script as this is generally
better practise. We have found that people end up running
the gate-check-commit script repeatedly instead of just
running a specified playbook.

This patch also adds the ceilometer containers to the
development stack diagram and rearranges a paragraph of text
which was in the wrong place.

Co-Authored-By: Matt Kassawara <mkassawara@gmail.com>
Change-Id: I020fa857dbafd72ffce7141616b61874dcf84047