Add galera_address to all playbooks installing galera_client
This patch includes the galera_address var for each playbook
that deploys a role depending on galera_client and thereby
dropping a .my.cnf onto the target.

Prior to this patch the .my.cnf would be dropped with the
galera_client's role default of 127.0.0.1 as the host address.

The address of 127.0.0.1 is intentionally left to deploy on
the galera servers so that the .my.cnf will be configured to
interact with the local database.

Change-Id: Ife33c68d825ecf1e6cdb6784802e9a099eb54289
Closes-Bug: #1487353