Set the correct variable in run-upgrade.sh
In run-upgrade.sh we are creating a temporary variables file that is
intended to be deleted post upgrade. We were setting the wrong variable
for the intended purpose - we want to allow containers to talk directly
to the repo containers since users will not be able to configure their
LB ahead of time.

By setting openstack_repo_url, we get the same effect and then we also
need to set the scheme (http://) appropriately so pip-link-build.py
appropriately sets up trusted-hosts in pip.conf

Closes-bug: 1482311
Change-Id: I47b2d7aabf669b529f874b1506430e82ad0506e6