Store used IPs from Juno config during upgrade
Kilo introduced a new structure to the networks for containers, and most
of the dynamic inventory script accounts for the differences between
Juno and Kilo. However, populating the used IP list happens prior to
restructuring the data on the first execution of Kilo code, which means
that the used IP processing is not looking for the right keys in the
host variables.

This results in non-deterministic IP address clashes, particularly with
repo containers, during upgrades. The IP addresses for a given network
are pulled from the pool of available IPs randomly.

Also, this change should only apply to Kilo; Liberty should not know
about Juno inventory details, since upgrades should only pass through
one version at a time.

Change-Id: Iae47e27cc7e926afe860e4896bf9f1ab291bc7c8
Closes-Bug: #1482375