Update kilo for new dev work - 10 Aug 2015
Update all repo SHAs for new work on 11.1.1

The python package pbr needs to be capped to <0.11
until https://review.openstack.org/211800 merges
in upstream heat.

Due to a build failure in the required python-openstacksdk
this patch temporarily fixes the version to one that builds
successfully. More information about this in:
 - https://launchpad.net/bugs/1484011

Related-Bug: #1484011
Closes-Bug: #1450733
Change-Id: I415e23f2d4e4f1770ac4a03313b0dd984db8a671