Set lxc.autodev=0 for cinder_volumes in containers
With cinder_volumes we were creating a specific udev device, this will
fail to mount if lxc.autodev=1. This should only be required when lvm is
a backend for the cinder_volumes container.

We can specify lxc.autodev=0 for cinder_volumes containers. To do this
properly we first check if lvm is in use. We can also use this to ensure
that redundant "lvm" configuration isn't setup on volumes hosts with no
lvm backend.

Additionally this will fix the formatting on the "udev" lxc.mount.entry
as it was adding additional spaces.

Change-Id: Iabe72003ebcfefe11d360131fdde64ca4b21a192
Closes-Bug: #1483650