Adds venv environment to tox for -infra docs job
The -infra templated job for docs requires that a venv be created using
tox. It will actually run this command to build the documentation:

  tox -evenv -- python setup.py build_sphinx

This patch also switches to using ansible 1.6.10 - it's the same version
as was present in requirements before, but the download will not be from
a Rackspace mirror. This is done to resolve a doc build failure when
the requirement is set using the previous method.

As there are no in-tree docs for the juno branch, this patch adds a bare
index which references the Rackspace documentation.

Change-Id: I69d7e0ab5613487c402f35d3ebb88915a48f50ce