Switch Nova/Tempest to use/test Cinder API v2
As per https://blueprints.launchpad.net/cinder/+spec/deprecate-v1-api
the Cinder API v1 was designated as deprecated in Juno and was meant
to be removed in Kilo, but appears to still be usable in Kilo.

This patch configures Nova to use the Cinder v2 API instead of the
v1 API.

This patch configures Tempest to test both the Cinder v1 and v2 API's
for now. Once the API is removed upstream we can change this.

DocImpact
Closes-Bug: #1481542
Change-Id: I6513c07e1b5d5653b84ad0b2b490ac8451f2bdf1
(cherry picked from commit 16d1a5150411eee9a9dd41e624d8b4e39fdf681c)