Set iptables-persistent install execution to append to log
This patch sets the command to install the iptables-persistent
package in the run-playbooks script to append its output to the
existing log instead of overwriting it.

This is to ensure that outputs from both commands are captured,
not just the second one.

Change-Id: I4d45382cb44c5f811818fe1229a63e6a2c2ecdee
Partial-Bug: #1485917
(cherry picked from commit a515e835bb5aef9cf4757ef3ac79f61ae560296a)