Enable/disable Swift/OpenStack deployment properly
This patch properly enables or disables Ceilometer, general
OpenStack and Swift deployment properly.

For the moment the containers will still be created, but
none of the related software will be installed, configured
or tested.

A fix to limit the containers created will be implemented
after a revision of the haproxy playbook/role to accommodate
this is implemented.

Closes-Bug: #1485945
Implements: blueprint split-aio-gates
Change-Id: Ia6657a02a6d1c53a4d76d7a17f74748ec9d2a2ee