Trigger restart after adding user to cephkeys
Currently, on an AIO, when you deploy w/ ceph support nova-compute
comes up non-functional and needs a restart.  Doing an strace on the
nova-compute process, you can see nova-compute is unable to read the
/etc/ceph/ceph.client.cinder.keyring file.  This reason for this is due
to the fact that cinder-volume and nova-compute both live on the same
host, and a bulk of the tasks in ceph_auth.yml (which do have valid
restarts) are being taken care of when the cinder installation playbook
is run.  When the nova installation playbook is run, ceph_auth.yml is
included after nova is installed but when nova is added to cephkeys
there is nothing else happening in that task file to trigger a restart
of nova.

Change-Id: I61e2709ce48ff80fd6eded9c1e6bd649bcd73088
Closes-Bug: #1486504