Allow a log dir that isn't the "service_name"
Galera logs by default to /var/log/mysql_logs these logs are being
ignored and not mounted properly on the host server, because the
service_name that is used to mount the log directory is "galera" and not
"mysql_logs".

This patch adds an option to set the log_directory specifically inside
the env.d yml files, but defaults to "service_name" when the
log_directory option isn't set. This will mean functionality remains
unchanged for all services.

Additionally, the log_directory has changed to "mysql_logs" for galera
containers, to resolve the galera log issue.

Change-Id: I18400248da455156dbba31b9c2ff254b58c51673
Closes-Bug: #1486955