Removed rpc-repo upstream pip deps
Changed variables such that the upstream rpc-repo is no longer used
as a source of truth for installing pip. While the repo is still
present as a default pip index (and it should remain to be as long
as RAX is maintaining the repo) it is now only used with the scope
of the repo server.

Change-Id: I524919dc9b7ea0f69ae00227969a6421bd64ec3b
Implements: blueprint remove-upstream-repo-dependency