Change AppArmor profile application order
This patch is a combination of two patches committed to master as the
first patch on its own results in continual gate check fails:

Patch 1:

Removed default lxc profile on container create

Having the lxc container create role drop the lxc-openstack apparmor
profile on all containers anytime its executed leads to the possibility
of the lxc container create task overwriting the running profile on a given
container. If this happens its likley to cause service interruption until the
correct profile is loaded for all containers its effected by the action.

To fix this issue the default "lxc-openstack" profile has been removed from the
lxc contianer create task and added to all plays that are known to be executed
within an lxc container. This will ensure that the profile is untouched on
subsequent runs of the lxc-container-create.yml play.

Closes-Bug: 1487130
(cherry picked from commit ffb701f8a3a325e0c321fb2d3e37eea25e66a8af)

Patch 2:

Wait for container ssh after apparmor profile update

This patch adds a wait for the container's sshd to be available
after the container's apparmor profile is updated. When the
profile is updated the container is restarted, so this wait is
essential to the success of the playbook's completion.

It also includes 3 retries which has been found to improve the
rate of success.

Due to an upstream change in behaviour with netaddr 0.7.16 we
need to pin the package to a lower version until Neutron is
adjusted and we bump the Neutron SHA.

Closes-Bug: #1490142
(cherry picked from commit a40cb5811822181369ee3269bc57d8bd19f05913)

Change-Id: Ifa4640be60c18f1232cc7c8b281fb1dfc0119e56