Add profiling for Ansible tasks
This should help find the slowest Ansible tasks that need to be optimized.

It is only enabled by default for gate jobs.

The profiling callback code is MIT licensed.  Per legal-discuss, I've added
the MIT license to the LICENSE.txt at the root of the repository along with
a note about which code is MIT licensed.

Thread on legal-discuss:

http://lists.openstack.org/pipermail/legal-discuss/2015-September/000398.html

Closes-Bug: #1488639

Change-Id: I46a2c495d9f8f1131ec08be99696efc39133f291